chore: refactor create page blocks
