+++ src/hooks/useImportBooks.tsx
@@ -7,6 +7,7 @@ import { renderTemplate } from "../utils/renderTemplate";
 import { createZettelId } from "../utils/zettelId";
 import { useLogseq } from "./useLogseq";
 
+
 export const nameToLink = ({ reverseNameOrder }: { reverseNameOrder?: boolean; } = {}) => (name: string): string => {
   let result = name;
 
@@ -50,13 +51,14 @@ const getPageByBookId = async (db: IDBProxy, bookId: string): Promise<PageEntity
 
 export async function getBookPage({ logseq, book, createPage = true }: GetBookPageParams) {
   const zettel = createZettelId();
-  let path = logseq.settings?.highlight_path ?? `highlights/{type}/{title}`;
-  path = renderTemplate(path, {
-    type: 'book',
-    title: book.title,
-    author: book.authors?.[0] ?? logseq.settings?.default_author ?? 'UnknownAuthor',
-    zettel,
-  });
+
+
+  const path = renderTemplate(logseq.settings?.highlight_path ?? `highlights/{type}/{title}`, {
+      type: 'book',
+      title: book.title,
+      author: book.authors?.[0] ?? logseq.settings?.default_author ?? 'UnknownAuthor',
+      zettel,
+    });
 
   console.info(`Loading path ${path}`);
   let page: PageEntity | null = null;
