+++ src/settingsSchema.ts
@@ -2,6 +2,14 @@ import { SettingSchemaDesc } from "@logseq/libs/dist/LSPlugin";
 
 export const SettingsSchema: Array<SettingSchemaDesc> = [
   {
+    title: 'Highlight Location',
+    key: 'highlight_location',
+    default: 'page',
+    type: 'enum',
+    enumChoices: ['page', 'journal'],
+    description: 'Should highlights be imported into a page for the book or into the journal and linked to the book?'
+  },
+  {
     title: 'Highlight Path',
     key: 'highlight_path',
     description: `Path where new highlights should be imported to.
